cast _sdc_sequence float as int

+++ target_bigquery.py
@@ -193,7 +193,7 @@ def persist_lines_job(project_id, dataset_id, lines=None, truncate=False, valida
             if add_tap_metadata:
                 msg.record['_sdc_version'] = msg.version
                 msg.record['_sdc_extracted_at'] = str(msg.time_extracted)
-                msg.record['_sdc_sequence'] = time.time() * 10000000000
+                msg.record['_sdc_sequence'] = int(time.time() * 10000000000)
 
             # NEWLINE_DELIMITED_JSON expects literal JSON formatted data, with a newline character splitting each row.
             dat = bytes(json.dumps(msg.record, cls=DecimalEncoder) + '\n', 'UTF-8')
@@ -288,7 +288,7 @@ def persist_lines_stream(project_id, dataset_id, lines=None, validate_records=Tr
             if add_tap_metadata:
                 msg.record['_sdc_version'] = msg.version
                 msg.record['_sdc_extracted_at'] = str(msg.time_extracted)
-                msg.record['_sdc_sequence'] = time.time() * 10000000000
+                msg.record['_sdc_sequence'] = int(time.time() * 10000000000)
 
             errors[msg.stream] = bigquery_client.insert_rows_json(tables[msg.stream], [msg.record])
             rows[msg.stream] += 1
